Less lines and easier to read/modify.

Practically no functional changes.