@arssher 2022-12-08
Based on [#277](https://github.com/neondatabase/neon/pull/2733) PR to less rebase it.