@arssher 2022-12-08
> Have you tried running build --timings to see how procmacros influence that?

I don't think it would be noticeable with other heavy deps, but let's see.

with patch
```