@arssher 2022-12-08
> Have you tried running build --timings to see how procmacros influence that?

I don't think it would be noticeable with other heavy deps, but let's see.

with patch
```
cargo build -p safekeeper --timings
Total time: | 466.4s (7m 46.4s)
  | safekeeper v0.1.0 bin "safekeeper" | 14.1s
```

without
```
total 483.2s (8m 3.2s)
safekeeper v0.1.0 bin "safekeeper" | 14.7s
```

and safekeeper v0.1.0 bin "safekeeper" doesn't have code generation part at all.