@petuhovskiy 2022-12-09
GitHub shows a large diff with `+1,225 −2,451`, but it seems that the patch is only a single commit. Not sure if this PR was needed at all, we could review it in the same neon-broker PR.